// src/app/shared/models/models.ts
export interface pageSelection {
  skip: number;
  limit: number;
}